Single socket implementation of per process power consumption
With this commit we begin to report per process power consumption. The
results are printed in microwatts.
To compute the power consumtion we find out the total socket power
consumtion using RAPL (considering only one socket so far) and the total
number of weighted cycles. We then assign power consumption to each
process using a proportional approach: the total power consumtion is
multiplied by the fraction (proc_cycles / total_cycles)