Use lazy loading to fetch info from k8s api
Since API requests are slow and might lead to unexpected crashes if the
response time is longer than the sleep time, we wrapped them inside
separate threads. The idea is to run a thread upon receiving the first
request and to return immediatly None, without waiting for the request
to be satisfied. Once we have the proper value we will update the
container_info accordingly.